Make custom predictor respond to probes when under load (#5)
* Make custom predictor respond to probes when under load

Python's base HTTP server is single threaded where it waits to accept the
next connection/request until the current one has been handled. In principle
for custom predictors this is what we want, to make sure the resources
requested by a custom predictor don't need to be shared between multiple
requests.

However, we also configure a readiness probe in Kubernetes where it tries
to connect to the custom predictor periodically without even making a request.
If the connection fails, the custom predictor is considered not ready and no
requests are routed to it. The problem now is that Python's `serve_forever()`
does the equivalent of:

```
while client := server_socket.accept():
    handle_request(client)
```

So we don't `accept()` the next connection until the current request has been
handled. This is in principle still fine because the operating system still
completes the TCP handshake, which is sufficient for the readiness probe to
succeed. However, the connection request from the readiness probe still takes
up a space in the operating system's buffer of unaccepted connections, and that
it stays there even after the remote site closes the connection again. By
default, Python [chooses this buffer size to be 5](https://docs.python.org/3/library/socketserver.html#socketserver.BaseServer.request_queue_size),
which is rather small. When the buffer is full, the OS waits for space to free
up before completing the TCP handshake, and this then causes the readiness probe
to fail. Since we probe every 5 seconds, it takes about 25 seconds for the
readiness probe to fail, which is lower than the 300 seconds of processing time
for custom predictors.

Instead of increasing the buffer size, in this PR I'm choosing a solution
where we buffer the incoming requests on the application level instead of
the OS level.